backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
ibm-watson==6.0.0
requests==2.32.3
pip freeze > requirements.txt
